Fix StrictDate and move KeyExpression
